refactor(env): unify index-array dtype on np.intp, remove int64 cast bridges
`bin_at`, `bins_in_region`, `boundary_bins`, and `_allocate_time_linear` now use
`np.intp` for their index arrays (pointer-sized: 64-bit on every platform,
including Windows, unlike `np.int_`), and `bin_at` normalizes divergent layout
index dtypes at the Environment boundary. This lets the two `.astype(np.int64)`
cast bridges in the occupancy path — previously needed because `bin_at` was
`np.int_` while `_allocate_time_linear` demanded `np.int64` — go away.

Trajectory-sequence arrays (`bin_sequence`, `BinSequenceWithRuns`) deliberately
keep their compact `int32` contract, and the `LayoutEngine.point_to_bin_index`
impls keep their internal dtypes below the boundary.

Occupancy output is byte-identical before/after (both allocation modes); mypy is
neutral on the default platform and `--platform win32`.

+++ src/neurospatial/environment/metrics.py
@@ -85,7 +85,7 @@ class EnvironmentMetrics:
 
     @cached_property
     @check_fitted  # Check only on first access, then value is cached
-    def boundary_bins(self: SelfEnv) -> NDArray[np.int_]:
+    def boundary_bins(self: SelfEnv) -> NDArray[np.intp]:
         """
         Identify boundary bins (bins on the edge of the active region).
 
@@ -95,7 +95,7 @@ class EnvironmentMetrics:
 
         Returns
         -------
-        NDArray[np.int_], shape (n_boundary_bins,)
+        NDArray[np.intp], shape (n_boundary_bins,)
             Array of bin IDs that are on the boundary.
 
         Notes
@@ -124,7 +124,7 @@ class EnvironmentMetrics:
             grid_shape=self.grid_shape,
             active_mask=self.active_mask,
             layout_kind=self._layout_type_used,
-        )
+        ).astype(np.intp, copy=False)
 
     @cached_property
     @check_fitted  # Check only on first access, then value is cached
